update d3d12 sdk and fix build

+++ library/modules/core/base/include/tiki/base/float32.hpp
@@ -72,7 +72,7 @@ namespace tiki
 
 		TIKI_FORCE_INLINE float mod( float value, float divisor )
 		{
-			return fmod( value, divisor );
+			return (float)fmod( value, divisor );
 		}
 
 		// return random value between 0 and 1
